requirements.txt
google-generativeai>=0.6.0
python-dotenv>=1.0.0